fixed type conflict in SDK_Sprite_AddAnimation, the mp3 library is now working

+++ lib_dev/include/sprite/SDK_sprite.h
@@ -110,7 +110,7 @@ SDK1_API int SDK_Sprite_AllocAnimation(SDK_Sprite *animated_sprite, uint16_t ani
     returns 0 for success, returns 1 for failure
     call SDL_GetError() for more info
 */
-SDK1_API int SDK_Sprite_AddAnimation(SDK_Sprite *animated_sprite, SDL_FRect src_rect, uint8_t amount_frames, double fps, double offset_width, uint16_t animation_index);
+SDK1_API int SDK_Sprite_AddAnimation(SDK_Sprite *animated_sprite, SDL_FRect src_rect, uint16_t amount_frames, double fps, double offset_width, uint16_t animation_index);
 
 /*
     Updates frames of an animation, must be called once per frame for animations to work
